move the `opentracing` package into the root dir of the repo
