+++ examples/dapperish.go
@@ -12,8 +12,8 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/opentracing/api-go/examples/dapperish"
-	"github.com/opentracing/api-go/opentracing"
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/examples/dapperish"
 )
 
 func client() {
